+++ src/rtcm3/index.ts
@@ -1,5 +1,6 @@
 export * from './decoder';
 export * from './ephemeris';
+export * from './legacy-obs';
 export * from './msm';
 export * from './station';
 export * from './stats';